Complete function for the user selection
The function prompts the user to enter either rock, paper, or scissors.
Leading and ending whitespace is trimmed and the letters are
all changed to lower case. Next the input is verified to be either
rock, paper, or scissors. If it is not, the user is notified and the
program prompts the user to enter their input again.