Write conditionals for player and computer wins
Added conditionals for a player win and a computer win.
Finished the game with a function rockPaperScissors which calls the
the game play function (fiveRounds), adds output based on the score of
each five round game, asks the user if they want to play again, and
provides a message if the user doesn't want to paly another game.