Create game code based on odin instructions
This is my version of rock paper scissors based on the suggestions
provided by TOP. The program has the same functionality that as the one
I created from scratch. I think that the for loop, instead of the
recursive calls, make it more readable. I also think that calling each
round separately is better style. All in all it was a good exercise and
shows that there are many ways to arrive at the right answer.